split code by category
