+++ modules/mex/src/cpp/MxCell.cpp
@@ -23,21 +23,52 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-if ~isbuiltin('mexPrintf')
-     status = copyfile('mexPrintf.c', tempdir());
-     assert_istrue(status);
-     cd(tempdir()); 
-     mex('mexPrintf.c');
-     run('loader.nls');
-end
+#include "mex.h"
+#include "matrix.h"
+#include "MxHelpers.hpp"
 //=============================================================================
-R = evalc('mexPrintf()');
-REF = 'Result 0
-Result 1
-
-ans =
-
-   Empty matrix : 1-by-0
-';
-assert_isequal(R, REF);
+mxArray*
+mxCreateCellArray(mwSize ndim, const mwSize* dims)
+{
+    return mxAllocateRealArray(ndim, dims, sizeof(void*), mxCELL_CLASS);
+}
+//=============================================================================
+mxArray*
+mxCreateCellMatrix(mwSize m, mwSize n)
+{
+    mwSize dims[2];
+    dims[0] = m;
+    dims[1] = n;
+    return mxCreateCellArray(2, dims);
+}
+//=============================================================================
+bool
+mxIsCell(const mxArray* pm)
+{
+    if (pm) {
+        return (pm->classID == mxCELL_CLASS);
+    }
+    return false;
+}
+//=============================================================================
+mxArray*
+mxGetCell(const mxArray* pm, mwIndex index)
+{
+    if (pm) {
+        return (((mxArray**)pm->realdata)[index]);
+    }
+    return nullptr;
+}
+//=============================================================================
+void
+mxSetCell(mxArray* pm, mwIndex index, mxArray* value)
+{
+    if (pm == nullptr) {
+        return;
+    }
+    if (!mxIsCell(pm)) {
+        return;
+    }
+    ((mxArray**)pm->realdata)[index] = value;
+}
 //=============================================================================
