+++ modules/mex/src/include/MxAsserts.h
@@ -23,21 +23,26 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-if ~isbuiltin('mexPrintf')
-     status = copyfile('mexPrintf.c', tempdir());
-     assert_istrue(status);
-     cd(tempdir()); 
-     mex('mexPrintf.c');
-     run('loader.nls');
-end
+#pragma once
 //=============================================================================
-R = evalc('mexPrintf()');
-REF = 'Result 0
-Result 1
-
-ans =
-
-   Empty matrix : 1-by-0
-';
-assert_isequal(R, REF);
+#include "nlsMex_exports.h"
+//=============================================================================
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+    //=============================================================================
+    NLSMEX_IMPEXP
+    void
+    mexPrintAssertion(const char* test, const char* fname, int linenum, const char* message);
+    //=============================================================================
+#define mxAssert(test, message)                                                                    \
+    ((test) ? (void)0 : mexPrintAssertion(#test, __FILE__, __LINE__, message))
+    //=============================================================================
+#define mxAssertS(test, message)                                                                   \
+    ((test) ? (void)0 : mexPrintAssertion("", __FILE__, __LINE__, message))
+    //=============================================================================
+#ifdef __cplusplus
+}
+#endif
 //=============================================================================
