+++ modules/mex/src/include/MxAttributes.h
@@ -23,21 +23,16 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-if ~isbuiltin('mexPrintf')
-     status = copyfile('mexPrintf.c', tempdir());
-     assert_istrue(status);
-     cd(tempdir()); 
-     mex('mexPrintf.c');
-     run('loader.nls');
-end
+#pragma once
 //=============================================================================
-R = evalc('mexPrintf()');
-REF = 'Result 0
-Result 1
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+    //=============================================================================
 
-ans =
-
-   Empty matrix : 1-by-0
-';
-assert_isequal(R, REF);
+    //=============================================================================
+#ifdef __cplusplus
+}
+#endif
 //=============================================================================
