+++ modules/mex/src/include/MxCharacter.h
@@ -23,21 +23,34 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-if ~isbuiltin('mexPrintf')
-     status = copyfile('mexPrintf.c', tempdir());
-     assert_istrue(status);
-     cd(tempdir()); 
-     mex('mexPrintf.c');
-     run('loader.nls');
-end
+#pragma once
 //=============================================================================
-R = evalc('mexPrintf()');
-REF = 'Result 0
-Result 1
+#include "nlsMex_exports.h"
+#include "mex.h"
+//=============================================================================
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+    //=============================================================================
 
-ans =
-
-   Empty matrix : 1-by-0
-';
-assert_isequal(R, REF);
+    NLSMEX_IMPEXP
+    int
+    mxGetString(const mxArray* pm, char* str, mwSize strlen);
+    //=============================================================================
+    NLSMEX_IMPEXP
+    mxArray*
+    mxCreateString(const char* str);
+    //=============================================================================
+    NLSMEX_IMPEXP
+    mxArray*
+    mxCreateCharMatrixFromStrings(mwSize m, const char** str);
+    //=============================================================================
+    NLSMEX_IMPEXP
+    mxArray*
+    mxCreateCharArray(mwSize ndim, const mwSize* dims);
+//=============================================================================
+#ifdef __cplusplus
+}
+#endif
 //=============================================================================
