+++ modules/mex/src/include/MxLogical.h
@@ -23,21 +23,29 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-if ~isbuiltin('mexPrintf')
-     status = copyfile('mexPrintf.c', tempdir());
-     assert_istrue(status);
-     cd(tempdir()); 
-     mex('mexPrintf.c');
-     run('loader.nls');
-end
+#pragma once
 //=============================================================================
-R = evalc('mexPrintf()');
-REF = 'Result 0
-Result 1
-
-ans =
-
-   Empty matrix : 1-by-0
-';
-assert_isequal(R, REF);
+#include "nlsMex_exports.h"
+#include "mex.h"
+//=============================================================================
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+    //=============================================================================
+    NLSMEX_IMPEXP
+    mxArray*
+    mxCreateLogicalScalar(mxLogical value);
+    //=============================================================================
+    NLSMEX_IMPEXP
+    mxArray*
+    mxCreateLogicalMatrix(mwSize m, mwSize n);
+    //=============================================================================
+    NLSMEX_IMPEXP
+    mxArray*
+    mxCreateLogicalArray(mwSize ndim, const mwSize* dims);
+    //=============================================================================
+#ifdef __cplusplus
+}
+#endif
 //=============================================================================
